backend/requirements.txt
# FastAPI Backend Dependencies
# Minimal version for development without database initially

# Core FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Authentication & Security (no cryptography for now)
python-multipart==0.0.12

# CORS & middleware
python-dotenv==1.0.1

# Data validation
email-validator==2.2.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Code quality
ruff==0.6.8  # Combined linter and formatter (faster than black+flake8)

# ================================================
# DATABASE DEPENDENCIES (Requires Build Tools)
# ================================================
# Uncomment these after installing Microsoft C++ Build Tools:
# https://visualstudio.microsoft.com/visual-cpp-build-tools/
#
# sqlalchemy==2.0.35
# asyncpg==0.29.0
psycopg2-binary==2.9.9
# alembic==1.13.3
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
